Server to server com WIP

+++ src/events/messageCreate/bridgedMessage.js
@@ -1,40 +1,57 @@
 const { WebhookClient } = require('discord.js');
 const { guildIsActive } = require('../../database/schemas/Guild');
+const { getRoomData } = require('../../database/schemas/Room');
 
 module.exports = async (client, message) => {
-  if (!message.guild) return;
-  if (message.author.bot) return;
-  if (!guildIsActive(message.guild.id)) return;
+  if (!message.guild || message.author.bot || !guildIsActive(message.guild.id)) return;
 
   const bridgeChannelName = 'crossover-channel';
   // Check if the message is in the designated channel
   if (message.channel.name === bridgeChannelName) {
-    const otherServerId =
-      message.guild.id === '828989782465839107'
-        ? '1189698827825987614'
-        : '828989782465839107';
-    const otherServer = client.guilds.cache.get(otherServerId);
-    if (!otherServer) return console.log('Other server not found');
-
-    const targetChannel = otherServer.channels.cache.find(
-      (channel) => channel.name === bridgeChannelName
-    );
-    if (!targetChannel) return console.log('Target channel not found');
-
-    // TODO: Recolect webhook ID and token for the servers and use them for communication after pairing the servers' token
-    const webhookClient = new WebhookClient({
-      id: '1212593463271493703',
-      token:
-        'H1EXTHTjs5LkmHHLLlAkEsRloyBFRNJcVWaEW5Y2ENaZ8zteybDtzyks60EBMTz6XVc6',
+    let data = await getRoomData(message.guild.id).catch((error) => {
+      console.error(`Error on getRoomData: ${error}`);
     });
-    webhookClient.send({
-      content: message.content,
-      username: message.author.username + ' from ' + message.guild.name,
-      avatarURL: message.author.displayAvatarURL(),
-      allowedMentions: { parse: [] },
-      target: targetChannel,
+    if (!data) {
+      message.channel.send('Room data not found');
+      console.warn('Room data not found');
+      return;
+    };
+
+    let roomParticipants = data.roomData.roomMembers;
+    let msgRecievers = roomParticipants.filter((id) => id !== message.guild.id);
+    if (!msgRecievers) {
+      message.channel.send('No other servers found. Invite them to the room to start chatting.');
+      console.warn('No other servers found');
+      return;
+    };
+
+    msgRecievers.forEach((id) => {
+      let otherServer = client.guilds.cache.get(id);
+      if (!otherServer) {
+        message.channel.send('Something happened while trying to send the message to the other server. Ensure the bot is in the server or that the server exists.');
+        console.warn('Server not found');
+        return;
+      };
+
+      let targetChannel = otherServer.channels.cache.find(
+        (channel) => channel.name === bridgeChannelName
+      );
+      if (!targetChannel) {
+        message.channel.send('The other server does not have a crossover channel');
+        console.warn('Crossover channel not found');
+        return;
+      };
+
+      const webhookClient = new WebhookClient({});
+      webhookClient.send({
+        content: message.content,
+        username: message.author.username + ' from ' + message.guild.name,
+        avatarURL: message.author.displayAvatarURL(),
+        allowedMentions: { parse: [] },
+        target: targetChannel,
+      });
+      webhookClient.destroy();
     });
-    webhookClient.destroy();
   } else {
     return;
   }
